3378 client - Workflow execution sheet/detail panel split
Workflow-execution-sheet detail panel split (WorkflowExecutionDetail +
useWorkflowExecutionDetail hook), WorkflowExecutionSheet /
WorkflowExecutionSheetWorkflowPanel updates (incl. embedded), and the
shared workflowExecutions queries.
