+++ client/src/shared/components/copilot/stores/useCopilotStore.ts
@@ -12,6 +12,7 @@ export enum MODE {
 }
 
 export enum Source {
+    WORKFLOW_EXECUTION = 'WORKFLOW_EXECUTION',
     WORKFLOW_EDITOR = 'WORKFLOW_EDITOR',
     CODE_EDITOR = 'CODE_EDITOR',
     CLUSTER_ELEMENT = 'CLUSTER_ELEMENT',
